Fix pkg Makefile not being honored on package rebuild
The build-packages.php script would step on changes implemented
by package specific Makefiles which are supported in other places
when it detected the changes files and forcibly rebuilt the pkg.
The Makefile would run in src/immutable-xyz then get steppd on in
src/pkgbuild/xyz.

This change had the rebuild process re-running the 'make install'
if the Makefile exists.

This was never relevant before because the only package using a
Makefile was sounds_us which doesn't ever really change.